backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.2.0
httpx==0.27.0
python-dotenv==1.0.0
requests==2.32.3
pytest==8.3.5
